Added custom abbreviations feature.
Abbreviations work with a file named "abbreviations.txt" where every line
is an abbreviation followed by its replacement.
If an argument in the command line matches an abbreviation, it will be replaced
by the items defined in the file.
For example, the file contains this line:
	sbs sandboxes

when the user types "dbdeployer sbs", it will be replaced with "dbdeployer sandboxes"

A more interesting example:
	group  replication --topology=group

Here, a command "dbdeployer group 8.0.4" becomes "dbdeployer replication --topology=group 8.0.4"
It is also possible to set variables in the replacement.
    sbdef --sandbox-directory={{.sb}} --port={{.port}}

To use this abbreviation, we need to provide the values for 'sb' and 'port'
dbdeployer sbdef:port=9000,sb=mysandbox single 8.0.4
it will become  "dbdeployer --sandbox-directory=mysandbox --port=9000 single 8.0.4

See the file abbreviations/sample_abbreviations.txt for some examples.