Fix Issue#34 "dbdeployer fails with .mylogin.cnf"
dbdeployer fails when a valid .mylogin.cnf is found in the $HOME
directory. The failure is due to the way .mylogin.cnf is designed: it
bypasses --no-defaults and --defaults-file (really funny, MySQL team!)
The only known way of using --no-defaults or --defaults-file with
.mylogin.cnf is to define an impossible path for it using
MYSQL_TEST_LOGIN_FILE environment variable.
dbdeployer does that, and also all the sandbox files use the same
workaround.