Fix --use-template to not require _template suffix
When using `dbdeployer deploy --use-template`, the current code will
accept either e.g. 'init_slaves' or 'init_slaves_template', but only the
latter will actually have an effect.

This is because findTemplate will return a match if
templateName+"_template" is a known template, but replaceTemplate still
uses the original templateName to store the updated template.